// src/app/models/asignacion.model.ts
export class AsignacionModel {
  LoteSolidoInicial: string;
  SolidoInicial: any[] = [];
  LoteLiquidoInicial: string;
  LiquidoInicial: any[] = [];
  LoteSolidoPrimaria: string;
  SolidoPrimaria: any[] = [];
  LoteLiquidoPrimaria: string;
  LiquidoPrimaria: any[] = [];
  LoteSolidoSegundaria: string;
  SolidoSegundaria: any[] = [];
  LoteLiquidoSegundaria: string;
  LiquidoSegundaria: any[] = [];
  IdMenu: string;
}